Continued to work on my NAV and finished my first row
I got the Nav setup finally that it is responsive properly. I have the brand on the on the left and the buttons on the right. For some reason but NAV buttons are jumping around when I hit a breakpoint. I have added my first row and I have designed it in a way that the images get a little smaller as the screen gets larger and so does the text. However when I set my screen to anything above 1200px I was expecting for all three columns to line up next to each other on the first row. I am going to start working on the second row and add a jumpbtron.