Jumbotron, NAV, Table, and adjusted the photos
I added a jumbotron because having the Nav fixed to the bottom leaves me wanting something on top of the three images, it felt plain. I also fixed the navbar at the bottom of the page and made sure that it was infront of everything on the Z axis. I also added a table with striped rows.